requirements.txt
pytest
pytest-bdd
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.5.2-cp37-cp37m-win_amd64.whl
allure-pytest
robotframework-python3
